WIP - making a treeview
I noticed that the built-in Outline view has a human-readable message whenever
it is empty, e.g., because the outline can't be created, or because an outline
doesn't make sense for the active editor.

If I create a TreeView and update in its `message` member based on the tree
data, I can get this behavior, too.